requirements.txt
google-generativeai>=0.4.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
pypdf>=4.0.0
pyyaml>=6.0.0